pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fts-diffusion"
version = "0.1.0"
description = "Paper-faithful implementation of FTS-Diffusion for financial time-series generation."
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
  "numpy>=1.26,<3",
  "pandas>=2.1,<3",
  "pyyaml>=6.0,<7",
  "scipy>=1.11,<2",
  "torch>=2.2,<3",
  "tqdm>=4.66,<5",
]

[project.scripts]
fts-train = "fts_diffusion.cli:train_main"
fts-sample = "fts_diffusion.cli:sample_main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

